First pass at a CUDA codepath. Only ExponentOp is implemented so far.
The obvious approach is simply to instantiate the Op objects on the
device, and write a kernel which loops through them and calls apply()
one pixel at a time. But many of the Ops have members of types that
aren't supported in device code (like std::string).

So I ended up creating a parallel class hierarchy of CudaOps. These
are stripped down to an apply() method and the members that go with
it, so they can exist on the device. CudaOps are created by
Op::getCudaOp(), which each subclass can override as necessary.

In ImageDesc, I added a new method: isDataOnCudaDevice(), and two new
subclasses, CudaPlanarImageDesc and CudaPackedImageDesc, which
override it to return true. In Processor::apply(), we check the
ImageDesc and switch to the CUDA path if necessary.

The CUDA path itself is just a single kernel. Each thread reads one
pixel from the image, applies all the ops, and stores the result.

I've tried to guard everything cuda specific with

== New Files ==
CudaSupport.cu/h: general-purpose support functions and macros for Cuda.
ProcessorCuda.cu: main implementation of the CUDA path
CudaOps.cuh: implementations of CUDA ops. This is #included directly
             in ProcessorCuda.cu because CUDA doesn't do separate
             compilation of device code.
CudaOps.h: factory functions for CudaOps.
ExponentOpsInternal.h: code shared by ExponentOp and CudaExponentOp

== Implementing CudaOps ==
This is the process I followed to implement CudaExponentOp:

1. Make a new file, ExponentOpsInternal.h, and move ApplyClampExponent
   there. This does the real work of ExponentOp::apply().
2. #include this file in ExponentOps.cpp and CudaOps.cuh so it's
   accessible to host and device code.
3. Implement CudaExponentOp class. It doesn't do much except call
   ApplyClampExponent in its apply() method.
4. Implement the factory function makeCudaExponentOp(). By hiding the
   CudaOps behind factories we can sidestep some compilation and
   linking issues.
5. Implement ExponentOp::getCudaOp(), which simply calls the factory.